Prepared basic ViewModel. Made basic service for movie search. Prepared display types for VC.

+++ Movie Picker/Models/Web/MovieResponse.swift
@@ -18,7 +18,7 @@ struct MovieResponse: Codable {
     let releaseDate: String?
     let title: String?
     let video: Bool?
-    let voteAverage: Int?
+    let voteAverage: Double?
     let voteCount: Int?
     
     enum CodingKeys: String, CodingKey {
